pyproject.toml
[project]
name = "quiz-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "google-generativeai",
    "python-dotenv>=1.0.0",
]
